# basic_Login/willow_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WillowCommerce API Example",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "https://willowcommerce-api.onrender.com"
    }
  ],
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "operationId": "HealthCheck",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        }
      }
    },

    "/orders/{tenant_id}/{order_id}": {
      "get": {
        "summary": "Get Order",
        "operationId": "GetOrder",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Order details",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },

    "/orders/{tenant_id}/{order_id}/cancel": {
      "post": {
        "summary": "Initiate Cancellation",
        "operationId": "CancelOrder",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Cancellation initiated",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },

    "/orders/{tenant_id}/{order_id}/replacement": {
      "post": {
        "summary": "Replacement Order",
        "operationId": "ReplacementOrder",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReplacementRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replacement initiated + label generated",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "502": {
            "description": "Label service failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },

    "/orders/{tenant_id}/{order_id}/return": {
      "post": {
        "summary": "Initiate Refund",
        "operationId": "RefundOrder",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RefundRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund initiated + label generated",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "502": {
            "description": "Label service failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },

    "/orders/{tenant_id}/{order_id}/humancontact": {
      "post": {
        "summary": "Human Contact",
        "operationId": "HumanContact",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          }
        ],
        "responses": {
          "200": {
            "description": "Human contact initiated",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          }
        }
      }
    },

    "/labels/{label_id}/view": {
      "get": {
        "summary": "View Label",
        "operationId": "ViewLabel",
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF stream",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "404": {
            "description": "Label not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },

    "/labels/{label_id}/download": {
      "get": {
        "summary": "Download Label",
        "operationId": "DownloadLabel",
        "parameters": [
          {
            "name": "label_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF download",
            "content": {
              "application/pdf": {
                "schema": { "type": "string", "format": "binary" }
              }
            }
          },
          "404": {
            "description": "Label not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },

    "/labels/{order_id}/{tenant_id}/get_label": {
      "get": {
        "summary": "Get Label Id",
        "operationId": "GetLabelId",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer", "format": "int64" }
          },
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Label lookup result",
            "content": {
              "application/json": {
                "schema": { "type": "object", "additionalProperties": true }
              }
            }
          },
          "404": {
            "description": "Label not found",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    }
  },

  "components": {
    "schemas": {
      "RefundRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": { "type": "string" }
        }
      },
      "ReplacementRequest": {
        "type": "object",
        "required": ["reason"],
        "properties": {
          "reason": { "type": "string" }
        }
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "detail": { "type": "string" }
        },
        "required": ["message"]
      }
    }
  }
}